+++ actions/src/sema4ai/actions/__init__.py
@@ -40,7 +40,7 @@ from ._protocols import IAction, Status
 from ._request import Request
 from ._secret import Secret
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 version_info = [int(x) for x in __version__.split(".")]